Changes: Added `/usr/lib` to `LD_LIBRARY_PATH` to fix ssl errors (I suspect partial flatpak library updates) Added the modified `LD_LIBRARY_PATH` to all shell/command calls Added code to obtain and maintain a copy of bwrap from the system, to work around permission issues
Known issues:
There is now an error about failing to connect the session bus. This appears to have no effect, sunshine launches without it and there is now other issues in the functions expected, so it seems the error can be safely ignored for now. Would still like to find why later.