Refactor build script and GitHub workflow / pipeline
The new version automatically updates the version number to the date of the merge of a PR into main / push to main. That way, when we want to push a new version into the store, we can simply use the latest commit and don't have to adjust the version manually.